+++ Mods ChoGGi/Fireworks Launcher/metadata.lua
@@ -4,21 +4,23 @@ return PlaceObj("ModDef", {
 	"steam_id", "1734717680",
 	"pops_any_uuid", "619213ed-1557-4bf5-843d-b33e0b8c5f48",
 	"lua_revision", 1007000, -- Picard
-	"version", 3,
+	"version", 4,
 	"version_major", 0,
-	"version_minor", 3,
-	"image", "Preview.png",
+	"version_minor", 4,
+	"image", "Preview.jpg",
 	"author", "ChoGGi",
 	"code", {
 		"Code/Script.lua",
 	},
+	"has_options", true,
 	"description", [[
 Adds a building to manually set off dome fireworks.
 
-Provides a minimal boost to terraforming heat (0.01) for a cost (100M).
-There is a visual only mod option.
+You can build it inside or outside (I would recommend inside).
 
-You can build it inside or outside, but outside will have some pretty particles.
+Green Planet bonus:
+Provides a minimal boost to terraforming heat (0.01) for a cost (100M), mod option to disable.
+Also an option to toggle rainfall.
 
 
 Requested by crowtalker.