+++ Mods ChoGGi/Show All Underground/items.lua
@@ -7,4 +7,10 @@ return {
 		"Help", T(302535920011793, "Disable mod without having to see missing mod msg."),
 		"DefaultValue", true,
 	}),
+	PlaceObj("ModItemOptionToggle", {
+		"name", "RevealDarkness",
+		"DisplayName", T(0000, "Reveal Darkness"),
+		"Help", T(0000, "Lights act more like surface."),
+		"DefaultValue", false,
+	}),
 }